+++ spring-cloud-stream-integration-tests/src/test/java/org/springframework/cloud/stream/config/ContentTypeOutboundSourceTests.java
@@ -16,6 +16,7 @@
 
 package org.springframework.cloud.stream.config;
 
+import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -38,6 +39,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * @author Ilayaperumal Gopinathan
+ * @author Vinicius Carvalho
  */
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringBootTest(classes = { ContentTypeOutboundSourceTests.TestSource.class })
@@ -53,12 +55,15 @@ public class ContentTypeOutboundSourceTests {
 	@Test
 	@SuppressWarnings("unchecked")
 	public void testMessageHeaderWhenNoExplicitContentTypeOnMessage() throws Exception {
-		testSource.output().send(MessageBuilder.withPayload("{\"message\":\"Hi\"}").build());
-		Message<String> received = (Message<String>) ((TestSupportBinder) binderFactory.getBinder(null,
+		testSource.output().send(MessageBuilder.withPayload("{\"message\":\"Hi\"}").setHeader(MessageHeaders.CONTENT_TYPE,"text/plain").build());
+		Message<?> received = ((TestSupportBinder) binderFactory.getBinder(null,
 				MessageChannel.class))
 						.messageCollector().forChannel(testSource.output()).poll();
-		assertThat(received.getHeaders().get(MessageHeaders.CONTENT_TYPE).toString()).isEqualTo("application/json");
-		assertThat(received).hasFieldOrPropertyWithValue("payload", "{\"message\":\"Hi\"}");
+		assertThat(received.getHeaders().get(MessageHeaders.CONTENT_TYPE).toString()).contains("text/plain");
+		Object payload = received.getPayload();
+		Assert.assertTrue(payload.getClass().isAssignableFrom(byte[].class));
+		byte[] contents = (byte[])payload;
+		Assert.assertEquals("{\"message\":\"Hi\"}", new String(contents));
 	}
 
 	@EnableBinding(Source.class)
