+++ spring-cloud-stream-integration-tests/src/test/java/org/springframework/cloud/stream/config/CustomHeaderPropagationTests.java
@@ -53,6 +53,11 @@ public class CustomHeaderPropagationTests {
 	private BinderFactory binderFactory;
 
 	@Test
+	/**
+	 * @since 2.0 The behavior of content type handling has changed.
+	 * All input/output channels have a default content type of application/json
+	 * When a processor or a source returns a String, and if the content type is json it will be quoted
+	 */
 	public void testCustomHeaderPropagation() throws Exception {
 		testProcessor.input().send(MessageBuilder.withPayload("{'name':'foo'}")
 				.setHeader(MessageHeaders.CONTENT_TYPE, "application/json")
@@ -65,8 +70,8 @@ public class CustomHeaderPropagationTests {
 		assertThat(received).isNotNull();
 		assertThat(received.getHeaders()).containsEntry("foo", "fooValue");
 		assertThat(received.getHeaders()).doesNotContainKey("bar");
-		assertThat(received.getHeaders()).doesNotContainKey(MessageHeaders.CONTENT_TYPE);
-		assertThat(received.getPayload()).isEqualTo("{'name':'foo'}");
+		assertThat(received.getHeaders()).containsKeys(MessageHeaders.CONTENT_TYPE);
+		assertThat(new String((byte[])received.getPayload())).isEqualTo("{'name':'foo'}");
 	}
 
 	@EnableBinding(Processor.class)
@@ -74,8 +79,9 @@ public class CustomHeaderPropagationTests {
 	public static class HeaderPropagationProcessor {
 
 		@ServiceActivator(inputChannel = "input", outputChannel = "output")
-		public String consume(String data) {
-			return data;
+		public Message<String> consume(String data) {
+			//if we don't force content to be String, it will be quoted on the outbound channel
+			return MessageBuilder.withPayload(data).setHeader(MessageHeaders.CONTENT_TYPE,"text/plain").build();
 		}
 
 	}
