+++ spring-cloud-stream-integration-tests/src/test/java/org/springframework/cloud/stream/config/CustomMessageConverterTests.java
@@ -28,6 +28,7 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.cloud.stream.annotation.Bindings;
 import org.springframework.cloud.stream.annotation.EnableBinding;
+import org.springframework.cloud.stream.annotation.StreamConverter;
 import org.springframework.cloud.stream.binder.BinderFactory;
 import org.springframework.cloud.stream.messaging.Source;
 import org.springframework.cloud.stream.test.binder.TestSupportBinder;
@@ -35,8 +36,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.integration.support.MessageBuilder;
-import org.springframework.integration.support.converter.ConfigurableCompositeMessageConverter;
-import org.springframework.integration.support.converter.DefaultDatatypeChannelMessageConverter;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageChannel;
 import org.springframework.messaging.MessageHeaders;
@@ -64,14 +63,14 @@ public class CustomMessageConverterTests {
 	private BinderFactory binderFactory;
 
 	@Autowired
+	@StreamConverter
 	private List<MessageConverter> customMessageConverters;
 
 	@Test
 	public void testCustomMessageConverter() throws Exception {
-		assertThat(customMessageConverters).hasSize(4);
+		assertThat(customMessageConverters).hasSize(2);
 		assertThat(customMessageConverters).extracting("class").contains(FooConverter.class,
-				BarConverter.class, DefaultDatatypeChannelMessageConverter.class,
-				ConfigurableCompositeMessageConverter.class);
+				BarConverter.class);
 		testSource.output().send(MessageBuilder.withPayload(new Foo("hi")).build());
 		@SuppressWarnings("unchecked")
 		Message<String> received = (Message<String>) ((TestSupportBinder) binderFactory.getBinder(null,
@@ -88,11 +87,13 @@ public class CustomMessageConverterTests {
 	public static class TestSource {
 
 		@Bean
+		@StreamConverter
 		public MessageConverter fooConverter() {
 			return new FooConverter();
 		}
 
 		@Bean
+		@StreamConverter
 		public MessageConverter barConverter() {
 			return new BarConverter();
 		}
