+++ spring-cloud-stream-integration-tests/src/test/java/org/springframework/cloud/stream/config/broken/DeserializeJSONToJavaTypeTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 the original author or authors.
+ * Copyright 2016 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.stream.config;
+package org.springframework.cloud.stream.config.broken;
 
+import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 import org.junit.Test;
@@ -25,26 +26,26 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.StreamListener;
 import org.springframework.cloud.stream.binder.BinderFactory;
 import org.springframework.cloud.stream.messaging.Processor;
 import org.springframework.cloud.stream.test.binder.TestSupportBinder;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.integration.annotation.ServiceActivator;
 import org.springframework.integration.support.MessageBuilder;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageChannel;
-import org.springframework.messaging.MessageHeaders;
-import org.springframework.messaging.handler.annotation.SendTo;
+import org.springframework.messaging.converter.MessageConverter;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * @author Marius Bogoevici
- * @since 1.2
  */
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringBootTest(classes = TextPlainToJsonConversionTest.FooProcessor.class, webEnvironment = SpringBootTest.WebEnvironment.NONE)
-public class TextPlainToJsonConversionTest {
+@SpringBootTest(classes = DeserializeJSONToJavaTypeTests.FooProcessor.class)
+public class DeserializeJSONToJavaTypeTests {
 
 	@Autowired
 	private Processor testProcessor;
@@ -52,45 +53,36 @@ public class TextPlainToJsonConversionTest {
 	@Autowired
 	private BinderFactory binderFactory;
 
-	@Test
-	public void testNoContentTypeToJsonConversionOnInput() throws Exception {
-		testProcessor.input().send(MessageBuilder.withPayload("{\"name\":\"Bar\"}").build());
-		Message<?> received = ((TestSupportBinder) binderFactory.getBinder(null, MessageChannel.class))
-				.messageCollector().forChannel(testProcessor.output()).poll(1, TimeUnit.SECONDS);
-		assertThat(received).isNotNull();
-		assertThat(((Foo) received.getPayload()).getName()).isEqualTo("transformed-Bar");
-	}
+	@Autowired
+	private List<MessageConverter> customMessageConverters;
 
 	@Test
-	public void testTextPlainToJsonConversionOnInput() throws Exception {
-		testProcessor.input().send(MessageBuilder.withPayload("{\"name\":\"Bar\"}")
-				.setHeader(MessageHeaders.CONTENT_TYPE, "text/plain").build());
+	public void testMessageDeserialized() throws Exception {
+		testProcessor.input().send(
+				MessageBuilder.withPayload("{\"name\":\"Bar\"}").setHeader("contentType", "application/json").build());
+		@SuppressWarnings("unchecked")
 		Message<?> received = ((TestSupportBinder) binderFactory.getBinder(null, MessageChannel.class))
 				.messageCollector().forChannel(testProcessor.output()).poll(1, TimeUnit.SECONDS);
 		assertThat(received).isNotNull();
-		assertThat(((Foo) received.getPayload()).getName()).isEqualTo("transformed-Bar");
+		assertThat(received.getPayload()).isInstanceOf(Foo.class);
+		assertThat((Foo) received.getPayload()).hasFieldOrPropertyWithValue("name", "Bar");
 	}
 
 	@EnableBinding(Processor.class)
 	@EnableAutoConfiguration
+	@PropertySource("classpath:/org/springframework/cloud/stream/config/fooprocesor/foo-sink.properties")
+	@Configuration
 	public static class FooProcessor {
 
-		@StreamListener("input")
-		@SendTo("output")
+		@ServiceActivator(inputChannel = "input", outputChannel = "output")
 		public Foo consume(Foo foo) {
-			Foo returnFoo = new Foo();
-			returnFoo.setName("transformed-" + foo.getName());
-			return returnFoo;
+			return foo;
 		}
-
 	}
 
 	public static class Foo {
 
 		private String name;
-
-		public Foo() {
-		}
 
 		public String getName() {
 			return name;
@@ -99,12 +91,5 @@ public class TextPlainToJsonConversionTest {
 		public void setName(String name) {
 			this.name = name;
 		}
-
-		@Override
-		public String toString() {
-			return "Foo{name='" + name + "'}";
-		}
-
 	}
-
 }
