+++ spring-cloud-stream-integration-tests/src/test/java/org/springframework/cloud/stream/config/broken/StreamListenerWithAnnotatedInputOutputArgsTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.stream.config;
+package org.springframework.cloud.stream.config.broken;
 
 import java.util.concurrent.TimeUnit;
