+++ spring-cloud-stream/src/main/java/org/springframework/cloud/stream/config/BindingProperties.java
@@ -54,7 +54,7 @@ public class BindingProperties {
 
 	// Properties for both inbound/outbound
 
-	private String contentType;
+	private String contentType = "application/json";
 
 	private String binder;
