+++ spring-cloud-stream/src/main/java/org/springframework/cloud/stream/converter/ObjectStringMessageConverter.java
@@ -65,10 +65,10 @@ public class ObjectStringMessageConverter extends AbstractMessageConverter {
 	protected Object convertToInternal(Object payload, MessageHeaders headers, Object conversionHint) {
 		if (payload != null) {
 			if ((payload instanceof byte[])) {
-				return new String((byte[]) payload, Charset.forName("UTF-8"));
+				return payload;
 			}
 			else {
-				return payload.toString();
+				return payload.toString().getBytes();
 			}
 		}
 		else {
